Migrating assertJsonCompatible() to new implementation.
